face-recognition-app/backend/face_recognition_service/requirements.txt
flask==3.0.2
flask-cors==6.0.0
facenet-pytorch==2.5.3
--find-links https://download.pytorch.org/whl/cpu
torch==2.7.0+cpu
torchvision==0.22.0+cpu
Pillow==11.2.1
numpy==2.2.6 